new clean code arch
