+++ handlers/stripe-disputes/src/dtos/index.ts
@@ -1,0 +1,4 @@
+// Barrel file for DTOs
+export * from './stripe-dispute-data.dto';
+export * from './listen-dispute-created.dto';
+export * from './listen-dispute-closed.dto';
