+++ handlers/stripe-disputes/src/handlers/index.ts
@@ -1,0 +1,3 @@
+// Barrel file for handlers
+export * from './dispute-created.handler';
+export * from './dispute-closed.handler';
