+++ handlers/stripe-disputes/src/interfaces/index.ts
@@ -1,0 +1,2 @@
+// Barrel file for interfaces
+export * from './payment-dispute.interface';
