+++ handlers/stripe-disputes/src/mappers/index.ts
@@ -1,0 +1,2 @@
+// Barrel file for mappers
+export * from './stripe-to-salesforce.mapper';
