+++ handlers/stripe-disputes/src/mappers/stripe-to-salesforce.mapper.ts
@@ -1,0 +1,35 @@
+import type { ListenDisputeCreatedRequestBody } from '../dtos';
+import type { PaymentDisputeRecord } from '../interfaces';
+import {
+	timestampToSalesforceDateTime,
+	timestampToSalesforceDate,
+} from '../helpers';
+
+/**
+ * Maps Stripe dispute webhook data to Salesforce Payment Dispute record format
+ */
+export function mapStripeDisputeToSalesforce(
+	stripeData: ListenDisputeCreatedRequestBody,
+): PaymentDisputeRecord {
+	const dispute = stripeData.data.object;
+
+	return {
+		attributes: {
+			type: 'Payment_Dispute__c',
+		},
+		Dispute_ID__c: dispute.id,
+		Charge_ID__c: dispute.charge,
+		Reason__c: dispute.reason,
+		Status__c: dispute.status,
+		Amount__c: dispute.amount / 100, // Convert cents to dollars
+		Evidence_Due_Date__c: timestampToSalesforceDate(
+			dispute.evidence_details.due_by,
+		),
+		Payment_Intent_ID__c: dispute.payment_intent,
+		Network_Reason_Code__c:
+			dispute.payment_method_details.card.network_reason_code,
+		Is_Charge_Refundable__c: dispute.is_charge_refundable,
+		Created_Date__c: timestampToSalesforceDateTime(dispute.created),
+		Has_Evidence__c: dispute.evidence_details.has_evidence,
+	};
+}
