+++ handlers/stripe-disputes/src/services/salesforceCreate.ts
@@ -1,23 +1,13 @@
-import { sfApiVersion } from '@modules/salesforce/src/config';
-import { z } from 'zod';
-import type { SalesforceAuthResponse } from './salesforceAuth';
-import type { PaymentDisputeRecord } from './stripeToSalesforceMapper';
-
-const SalesforceCreateErrorSchema = z.object({
-	statusCode: z.string().optional(),
-	message: z.string(),
-	fields: z.array(z.string()),
-});
-
-const SalesforceCreateResponseSchema = z.object({
-	id: z.string().optional(),
-	success: z.boolean(),
-	errors: z.array(SalesforceCreateErrorSchema),
-});
-
-export type SalesforceCreateResponse = z.infer<
-	typeof SalesforceCreateResponseSchema
->;
+import {
+	buildSalesforceUpsertUrl,
+	buildSalesforceUpsertOptions,
+} from '../helpers';
+import { SalesforceCreateResponseSchema } from '../zod-schemas';
+import type {
+	SalesforceAuthResponse,
+	SalesforceCreateResponse,
+} from '../types';
+import type { PaymentDisputeRecord } from '../interfaces';
 
 /**
  * Upserts a Payment Dispute record in Salesforce using Dispute_ID__c as external ID
@@ -31,19 +21,15 @@ export async function upsertPaymentDisputeInSalesforce(
 
 	try {
 		// Use PATCH with external ID for upsert: /sobjects/SObjectType/FieldName/FieldValue
-		const url = `${authResponse.instance_url}/services/data/${sfApiVersion()}/sobjects/Payment_Dispute__c/Dispute_ID__c/${paymentDispute.Dispute_ID__c}`;
+		const url = buildSalesforceUpsertUrl(authResponse, paymentDispute);
 
 		// Remove Dispute_ID__c from the body since it's in the URL
 		const { Dispute_ID__c, ...paymentDisputeBody } = paymentDispute;
 
-		const options = {
-			method: 'PATCH',
-			headers: {
-				Authorization: `Bearer ${authResponse.access_token}`,
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify(paymentDisputeBody),
-		};
+		const options = buildSalesforceUpsertOptions(
+			authResponse,
+			paymentDisputeBody,
+		);
 
 		const response = await fetch(url, options);
