+++ handlers/stripe-disputes/src/types/index.ts
@@ -1,0 +1,2 @@
+// Barrel file for types
+export * from './salesforce.types';
