+++ handlers/stripe-disputes/src/zod-schemas/salesforce-auth.schema.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+
+export const SalesforceAuthResponseSchema = z.object({
+	access_token: z.string(),
+	instance_url: z.string().url(),
+	id: z.string().url(),
+	token_type: z.string(),
+	issued_at: z.string(),
+	signature: z.string(),
+});
